backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
pymysql>=1.1.0,<2.0.0
oss2>=2.18.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
python-multipart>=0.0.9,<1.0.0
python-socketio>=5.10.0,<6.0.0
requests>=2.31.0,<3.0.0
opencv-python>=4.8.0
numpy>=1.24.0,<2.0.0
onnxruntime>=1.16.0
Pillow>=10.0.0
supervision>=0.19.0
filterpy>=1.4.5
scipy>=1.10.0
matplotlib>=3.7.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0
httpx>=0.27.0
